requirements.txt
boto3
botocore
keras==2.15.0
tensorflow==2.15.0
pillow
numpy
fastapi==0.78.0
uvicorn==0.18.3
streamlit